# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "AlphaMaster API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/admin/media/{media}": {
            "get": {
                "operationId": "admin.media.show",
                "summary": "Inspect one record",
                "tags": [
                    "MediaAdmin"
                ],
                "parameters": [
                    {
                        "name": "media",
                        "in": "path",
                        "required": true,
                        "description": "The media ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MediaAdminResource"
                                        },
                                        "meta": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "delete": {
                "operationId": "admin.media.destroy",
                "summary": "Soft delete. The bytes are purged later by the retention job, so a mistaken\nremoval here is recoverable",
                "tags": [
                    "MediaAdmin"
                ],
                "parameters": [
                    {
                        "name": "media",
                        "in": "path",
                        "required": true,
                        "description": "The media ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "meta": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/admin/roles/{role}": {
            "put": {
                "operationId": "admin.roles.update",
                "summary": "Replace a role's label and permissions",
                "tags": [
                    "RoleAdmin"
                ],
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "description": "The role ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/RoleResource"
                                        },
                                        "meta": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "admin.roles.destroy",
                "summary": "Delete a role",
                "tags": [
                    "RoleAdmin"
                ],
                "parameters": [
                    {
                        "name": "role",
                        "in": "path",
                        "required": true,
                        "description": "The role ID",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "null"
                                        },
                                        "meta": {
                                            "type": "string"
                                        }
                                    },
                                    "required": [
                                        "success",
                                        "message",
                                        "data",
                                        "meta"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "MediaAdminResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string"
                    },
                    "collection": {
                        "type": "string"
                    },
                    "original_filename": {
                        "type": "string"
                    },
                    "mime_type": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "type_label": {
                        "type": "string"
                    },
                    "size_bytes": {
                        "type": "integer"
                    },
                    "checksum": {
                        "type": "string"
                    },
                    "visibility": {
                        "type": "string"
                    },
                    "visibility_label": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "status_label": {
                        "type": "string"
                    },
                    "scan_status": {
                        "type": "string"
                    },
                    "scan_status_label": {
                        "type": "string"
                    },
                    "failure_reason": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "attachable_type": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "attachable_id": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "uploaded_by": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "collection",
                    "original_filename",
                    "mime_type",
                    "type",
                    "type_label",
                    "size_bytes",
                    "checksum",
                    "visibility",
                    "visibility_label",
                    "status",
                    "status_label",
                    "scan_status",
                    "scan_status_label",
                    "failure_reason",
                    "attachable_type",
                    "attachable_id",
                    "uploaded_by",
                    "created_at"
                ],
                "title": "MediaAdminResource"
            },
...
            "RoleRequest": {
                "type": "object",
                "properties": {
                    "label": {
                        "type": "string",
                        "description": "`label` is what an administrator types, and it is named for what it\nis: the response's `name` is the machine identifier, so reusing that\nword here would give one field two meanings across the same\nresource. The identifier is derived from this server-side and is\nimmutable afterwards (ADR 0029 item 12), so this carries no\nidentifier grammar and no uniqueness rule \u2014 two roles may read the\nsame in a list while remaining distinct underneath.",
                        "maxLength": 100
                    },
                    "permissions": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "description": "Only catalogued permissions may be attached, so a role cannot be given\na permission string the platform does not actually enforce anywhere.",
                            "enum": [
                                "users.view",
                                "users.create",
                                "users.update",
                                "users.delete",
                                "settings.view",
                                "settings.update",
                                "settings.rollback",
                                "settings.security.update",
                                "settings.secrets.manage",
                                "audit.view",
                                "audit.manage",
                                "settings.backup.manage",
                                "roles.view",
                                "roles.update",
                                "permissions.view",
                                "permissions.update",
                                "integrations.view",
                                "integrations.update",
                                "notifications.view",
                                "notifications.update",
                                "media.view",
                                "media.delete"
                            ]
                        },
                        "maxItems": 100
                    }
                },
                "required": [
                    "label",
                    "permissions"
                ],
                "title": "RoleRequest"
            },
            "RoleResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string",
                        "description": "`name` is the immutable machine identifier and does not move.\n`name_label` sits beside it (ADR 0031) so a client never has to\ninfer a human label from the identifier."
                    },
                    "name_label": {
                        "type": "string"
                    },
                    "permissions": {
                        "type": "object",
                        "additionalProperties": {}
                    }
                },
                "required": [
                    "id",
                    "name",
                    "name_label",
                    "permissions"
                ],
                "title": "RoleResource"
            },
...
        },
        "responses": {
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}